+++ log/output-react/src/components/row/components/step/StepToggle.tsx
@@ -1,8 +1,8 @@
-import { FC, useCallback, MouseEvent } from 'react';
+import { FC, useCallback, MouseEvent, KeyboardEvent } from 'react';
 import { Box } from '@robocorp/components';
 import { styled } from '@robocorp/theme';
 
-import { Entry, Type } from '~/lib/types';
+import { Entry } from '~/lib/types';
 import { useLogContext } from '~/lib';
 
 type Props = {
@@ -25,14 +25,27 @@ const Button = styled.button`
 `;
 
 export const StepToggle: FC<Props> = ({ entry }) => {
-  const { filteredEntries, isExpanded, toggleEntry } = useLogContext();
+  const { filteredEntries, isExpanded, toggleEntryExpandState } = useLogContext();
 
   const expanded = isExpanded(entry.id);
 
-  const onToggle = useCallback(
+  const onClickToggleExpand = useCallback(
     (e: MouseEvent) => {
       e.stopPropagation();
-      toggleEntry(entry.id);
+      toggleEntryExpandState(entry.id);
+    },
+    [entry.id],
+  );
+
+  const onKeyDownIgnore = useCallback(
+    (event: KeyboardEvent) => {
+      if (event.key === 'Enter') {
+        // When the button has focus and the user presses enter, we'll automatically have
+        // an onClick event (so, that's where the onClick will toggle the entry), but also
+        // we don't want the parent to have the same event (as that'd make the details
+        // be seen for the item), so, we have to stop the propagation of this event.
+        event.stopPropagation();
+      }
     },
     [entry.id],
   );
@@ -43,7 +56,12 @@ export const StepToggle: FC<Props> = ({ entry }) => {
   }
 
   return (
-    <Button onClick={onToggle} aria-label="Toggle item" className="toggleExpand">
+    <Button
+      onClick={onClickToggleExpand}
+      onKeyDown={onKeyDownIgnore}
+      aria-label="Toggle item"
+      className="toggleExpand"
+    >
       {expanded ? (
         <svg width="12" height="12" fill="none" xmlns="http://www.w3.org/2000/svg">
           <path d="M10.594 3.959A.75.75 0 0 0 10 2.75H1.5a.75.75 0 0 0-.593 1.209l4.25 5.5a.75.75 0 0 0 1.186 0l4.25-5.5Z" />